ai-processing/resume_optimizer/requirements.txt
openai>=0.27.0
numpy>=1.21.0
scikit-learn>=1.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
torch>=1.9.0,<2.0.0
transformers>=4.15.0
spacy>=3.0.0
tqdm>=4.62.0
pandas>=1.3.0
python-dotenv>=0.19.0 